OD-11189 Fix compilation errors
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@48796 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ checkout-api/src/main/groovy/ish/oncourse/willow/checkout/persistent/CreateWaitingList.groovy
@@ -28,7 +28,7 @@ class CreateWaitingList {
         waitingList.student = contact.student
         waitingList.course = new GetCourse(context, college, w.courseId).get()
         waitingList.detail = w.detail
-        waitingList.potentialStudents = w.studentsCount
+        waitingList.potentialStudents = w.studentsCount.intValue()
 
         (w.fieldHeadings.fields.flatten() as List<Field>).each { f  ->
             String value = StringUtils.trimToNull(f.value)
